backend/pyproject.toml
[project]
name = "blogpost-backend"
version = "0.1.0"
description = "BlogPost API built with Django REST Framework"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [
  { name = "Emilio Sala" }
]
keywords = ["django", "drf", "rest", "api", "blog"]

dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15,<4.0",
  "djangorestframework-simplejwt>=5.3,<6.0",
  "psycopg[binary]>=3.1,<4.0",
  "django-cors-headers>=4.3,<5.0",
  "python-dotenv>=1.0,<2.0",
  "django-allauth>=65.11.2",
  "requests>=2.32.5",
  "Pillow>=10.0.0,<11.0.0",
  "django-filter>=24.0,<25.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-django>=4.8,<5.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5.0,<1.0",
  "mypy>=1.10,<2.0",
]

[tool.uv]
package = false 

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"


